server/requirements.txt
Flask
Flask-CORS
pydub
openai-whisper
transformers
torch
requests
